Rename files and add has_pharmacy_first group_by to measures
